refactor: extract timer utils
