+++ src/app/components/SwapComponent.tsx
@@ -58,6 +58,7 @@ import {
   getGnosisPublicClient,
   setGnosisRpcUrl,
 } from "./utils";
+import { useTimer } from "./TimerUtils";
 
 import { getGnosisQuote, getCrossChainQuote } from "./CustomQuotes";
 import { handleFileUpload as uploadFile, isArchiveFile } from "./FileUploadUtils";
@@ -137,12 +138,7 @@ const SwapComponent: React.FC = () => {
 
   const [activeDropdown, setActiveDropdown] = useState<string | null>(null);
 
-  const [estimatedTime, setEstimatedTime] = useState<number | null>(null);
-  const [remainingTime, setRemainingTime] = useState<number | null>(null);
-
   const [serveUncompressed, setServeUncompressed] = useState(true);
-
-  const timerIntervalRef = useRef<NodeJS.Timeout | null>(null);
 
   // Add a ref to track the current wallet client
   const currentWalletClientRef = useRef(walletClient);
@@ -151,6 +147,8 @@ const SwapComponent: React.FC = () => {
   useEffect(() => {
     currentWalletClientRef.current = walletClient;
   }, [walletClient]);
+
+  const { estimatedTime, setEstimatedTime, remainingTime, formatTime, resetTimer } = useTimer(statusMessage);
 
   // Add a ref for the abort controller
   const priceEstimateAbortControllerRef = useRef<AbortController | null>(null);
@@ -1145,71 +1143,6 @@ const SwapComponent: React.FC = () => {
 
   const handleOpenDropdown = (dropdownName: string) => {
     setActiveDropdown(dropdownName);
-  };
-
-  useEffect(() => {
-    // Clear any existing timer first
-    if (timerIntervalRef.current) {
-      clearInterval(timerIntervalRef.current);
-      timerIntervalRef.current = null;
-    }
-
-    // Start a new timer if we have an estimated time and we're in the Route step
-    if (estimatedTime !== null && statusMessage.step === "Route") {
-      console.log("Starting timer with duration:", estimatedTime);
-
-      // Initialize the remaining time if it's not set
-      if (remainingTime === null) {
-        const buffer: number = 10;
-        setRemainingTime(estimatedTime + buffer);
-      }
-
-      // Create the interval
-      timerIntervalRef.current = setInterval(() => {
-        setRemainingTime((prevTime) => {
-          const newTime = prevTime !== null ? prevTime - 1 : 0;
-          // Reduce log frequency to avoid console spam
-          if (newTime % 5 === 0) {
-            console.log("Timer tick, remaining time:", newTime);
-          }
-
-          if (newTime <= 0) {
-            if (timerIntervalRef.current) {
-              clearInterval(timerIntervalRef.current);
-              timerIntervalRef.current = null;
-            }
-            return 0;
-          }
-          return newTime;
-        });
-      }, 1000);
-    }
-
-    // Clean up function
-    return () => {
-      if (timerIntervalRef.current) {
-        console.log("Cleaning up timer");
-        clearInterval(timerIntervalRef.current);
-        timerIntervalRef.current = null;
-      }
-    };
-  }, [estimatedTime, statusMessage.step]);
-
-  const formatTime = (seconds: number): string => {
-    if (seconds <= 0) return "0:00";
-    const minutes = Math.floor(seconds / 60);
-    const remainingSeconds = seconds % 60;
-    return `${minutes}:${remainingSeconds.toString().padStart(2, "0")}`;
-  };
-
-  // 3. Update the reset function to also clear the interval
-  const resetTimer = () => {
-    if (timerIntervalRef.current) {
-      clearInterval(timerIntervalRef.current);
-      timerIntervalRef.current = null;
-    }
-    setEstimatedTime(null);
-    setRemainingTime(null);
   };
 
   // Reset insufficientFunds whenever the selected token changes
